+++ SplitwiseApp.Core/ApiControllers/FriendsController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using SplitwiseApp.DomainModels.Models;
 using SplitwiseApp.Repository.DTOs;
-using SplitwiseApp.Repository.Friends;
+using SplitwiseApp.Repository.Friend;
+using SplitwiseApp.Repository.User;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -13,41 +15,55 @@ namespace SplitwiseApp.Core.ApiControllers
     public class FriendsController : ControllerBase
     {
         private readonly IFriends _friends;
-        public FriendsController(IFriends friends)
+        private readonly IUser _user;
+        public FriendsController(IFriends friends,IUser user)
         {
             _friends = friends;
+            _user = user;
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<IEnumerable<FriendsDTO>>> GetFriends(string id)
         {
-
-            IEnumerable<FriendsDTO> friendDtos = await _friends.GetFriends(id);
-            return Ok(friendDtos);
+            if (_user.UserExists(id))
+            {
+                IEnumerable<FriendsDTO> friendDtos = await _friends.GetFriends(id);
+                return Ok(friendDtos);
+            }
+            return NotFound();
         }
 
         [HttpPost]
-        public async Task<ActionResult<FriendsDTO>> AddAFriend(FriendsDTO friends)
+        public IActionResult AddAFriend(Friends friends)
         {
-
-            FriendsDTO addFriend = await _friends.AddAFriend(friends);
-            return Ok(addFriend);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+            _friends.AddAFriend(friends);
+            return Ok();
         }
 
         [HttpPut]
-        public async Task<ActionResult<FriendsDTO>> UpdateFriend(FriendsDTO friends)
+        public IActionResult UpdateFriend(Friends friends)
         {
-
-            FriendsDTO updateFriend = await _friends.UpdateAFriend(friends);
-            return Ok(updateFriend);
+            if (_friends.FriendExist(friends.friendId))
+            {
+                _friends.UpdateAFriend(friends);
+                return Ok();
+            }
+            return NotFound();
         }
 
         [HttpDelete]
-        public async Task<ActionResult<FriendsDTO>> DeleteFriend(int id)
+        public IActionResult DeleteFriend(int id)
         {
-
-            FriendsDTO deleteFriend = await _friends.DeleteAFriend(id);
-            return Ok(deleteFriend);
+            if (_friends.FriendExist(id))
+            {
+                _friends.DeleteAFriend(id);
+                return Ok();
+            }
+            return NotFound();
         }
     }
 }
