+++ SplitwiseApp.Core/ApiControllers/GroupsController.cs
@@ -4,8 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using SplitwiseApp.DomainModels.Models;
 using SplitwiseApp.Repository.DTOs;
-using SplitwiseApp.Repository.Groups;
+using SplitwiseApp.Repository.Group;
+using SplitwiseApp.Repository.User;
 
 namespace SplitwiseApp.Core.ApiControllers
 {
@@ -14,52 +16,70 @@ namespace SplitwiseApp.Core.ApiControllers
     public class GroupsController : ControllerBase
     {
         private readonly IGroups _groups;
-        public GroupsController(IGroups groups)
+        private readonly IUser _user;
+        public GroupsController(IGroups groups,IUser users)
         {
             _groups = groups;
+            _user = users;
 
         }
         // GET: api/Groups
         [HttpGet("{id}")]
+        [Route("{userid}")]
         public async Task<ActionResult<IEnumerable<GroupsDTO>>> GetGroupForAUser(string userId)
         {
-
-            IEnumerable<GroupsDTO> groupUser = await _groups.GetGroupByUserId(userId);
-            return Ok(groupUser);
+            if (_user.UserExists(userId))
+            {
+                IEnumerable<GroupsDTO> groupUser = await _groups.GetGroupByUserId(userId);
+                return Ok(groupUser);
+            }
+            return NotFound();
         }
 
         // GET: api/Groups
         [HttpGet("{id}")]
         public async Task<ActionResult<GroupsDTO>> GetGroup(int groupId)
         {
-
-            GroupsDTO groupsDtos = await _groups.GetGroupByGroupId(groupId);
-            return Ok(groupsDtos);
+            if (_groups.GroupExist(groupId))
+            {
+                GroupsDTO groupsDtos = await _groups.GetGroupByGroupId(groupId);
+                return Ok(groupsDtos);
+            }
+            return NotFound();
         }
 
         // POST: api/Groups
         [HttpPost]
-        public async Task<ActionResult<GroupsDTO>> AddAGroup(GroupsDTO groups)
+        public IActionResult AddAGroup(Groups groups)
         {
-
-            GroupsDTO addGroup = await _groups.AddGroupForUser(groups);
-            return Ok(addGroup);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+             _groups.AddGroupForUser(groups);
+            return Ok();
         }
 
         [HttpPut]
-        public async Task<ActionResult<GroupsDTO>> UpdateGroup(GroupsDTO groups)
+        public IActionResult UpdateGroup(Groups groups)
         {
-
-            GroupsDTO updateGroup = await _groups.UpdateAGroup(groups);
-            return Ok(updateGroup);
+            if (_groups.GroupExist(groups.groupId))
+            {
+                _groups.UpdateAGroup(groups);
+                return Ok();
+            }
+            return NotFound();
         }
 
         [HttpDelete]
-        public async Task<ActionResult<GroupsDTO>> DeleteGroup(int groupId)
+        public IActionResult DeleteGroup(int groupId)
         {
-
-            GroupsDTO deleteGroup = await _groups.DeleteAGroupById(groupId);
-            return Ok(deleteGroup);
+            if (_groups.GroupExist(groupId))
+            {
+                _groups.DeleteAGroupById(groupId);
+                return Ok();
+            }
+            return NotFound();
         }
     }
 }
